Move and RESTify images
