+++ torchweb/views.py
@@ -13,6 +13,8 @@ from collections import (
     WorkflowSettings,
 )
 
+from torch.collections.image import upload_images
+
 views = Blueprint("views", __name__)
 
 
@@ -95,86 +97,3 @@ def overview():
     return render_template("overview.html", user=current_user)
 
 
-@views.route("/history", methods=["GET", "POST"])
-@views.route("/delete-institution", methods=["POST"])
-def delete_institution():
-    institution = json.loads(request.data)
-    institutionId = institution["institutionId"]
-    institution = Institution.query.get(institutionId)
-    if institution:
-        db.session.delete(institution)
-        db.session.commit()
-
-    return jsonify({})
-
-
-@views.route("/upload", methods=["POST"])
-def upload():
-    """Handle the upload of a file."""
-    form = request.form
-
-    # Create a unique "session ID" for this particular batch of uploads.
-    upload_key = str(uuid4())
-
-    # Is the upload using Ajax, or a direct POST by the form?
-    is_ajax = False
-    if form.get("__ajax", None) == "true":
-        is_ajax = True
-
-    # Target folder for these uploads.
-    target = "webapp/static/uploads/{}".format(upload_key)
-    try:
-        os.mkdir(target)
-    except:
-        if is_ajax:
-            return ajax_response(
-                False, "Couldn't create upload directory: {}".format(target)
-            )
-        else:
-            return "Couldn't create upload directory: {}".format(target)
-
-    print("=== Form Data ===")
-    for key, value in list(form.items()):
-        print(key, "=>", value)
-
-    for upload in request.files.getlist("file"):
-        filename = upload.filename.rsplit("/")[0]
-        destination = "/".join([target, filename])
-        print("Accept incoming file:", filename)
-        print("Save it to:", destination)
-        upload.save(destination)
-
-    if is_ajax:
-        return ajax_response(True, upload_key)
-    else:
-        flash("Upload completed!", category="success")
-
-
-@views.route("/files/<uuid>")
-def upload_complete(uuid):
-
-    # Get their files.
-    root = "webapp/static/uploads/{}".format(uuid)
-    if not os.path.isdir(root):
-        return "Error: UUID not found!"
-
-    files = []
-    for file in glob.glob("{}/*.*".format(root)):
-        fname = file.split(os.sep)[-1]
-        files.append(fname)
-
-    return render_template(
-        "files.html",
-        uuid=uuid,
-        files=files,
-    )
-
-
-def ajax_response(status, msg):
-    status_code = "ok" if status else "error"
-    return json.dumps(
-        dict(
-            status=status_code,
-            msg=msg,
-        )
-    )
